refactor : LadderLine 생성 메서드 리팩토링
