+++ src/main/java/nextstep/ladder/LadderGame.java
@@ -1,17 +1,14 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.line.LadderLine;
+import nextstep.ladder.domain.line.LadderLines;
 import nextstep.ladder.domain.line.LadderLineGenerator;
 import nextstep.ladder.domain.point.RightPointRandomStrategy;
 import nextstep.ladder.domain.result.LadderResult;
 import nextstep.ladder.domain.result.Results;
-import nextstep.ladder.domain.user.User;
 import nextstep.ladder.domain.user.UserGenerator;
 import nextstep.ladder.domain.user.Users;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
-
-import java.util.Map;
 
 public class LadderGame {
 
@@ -24,12 +21,12 @@ public class LadderGame {
         Results results = new Results(InputView.inputLadderResult());
 
         int height = InputView.inputLadderHeight();
-        LadderLine ladderLine = LadderLineGenerator.generateLadderLine(height, countOfUser, new RightPointRandomStrategy());
+        LadderLines ladderLines = LadderLineGenerator.generateLadderLine(height, countOfUser, new RightPointRandomStrategy());
 
         LadderResult ladderResults = new LadderResult(users.getUsers());
-        ladderLine.goDownALadder(ladderResults.getLadderResult());
+        ladderLines.goDownALadder(ladderResults.getLadderResult());
 
-        ResultView.printLadderResult(users.getUsers(), ladderLine, results.getResults());
+        ResultView.printLadderResult(users.getUsers(), ladderLines, results.getResults());
 
         ResultView.printGameResult(ladderResults.getLadderResult(), results.getResults());
     }
