+++ src/main/java/nextstep/ladder/domain/line/LadderLineGenerator.java
@@ -9,9 +9,9 @@ import static nextstep.ladder.utils.CommonConstant.NUMBER_ZERO;
 
 public class LadderLineGenerator {
 
-    public static LadderLine generateLadderLine(int height, int countOfUser, RightPointStrategy rightPointStrategy) {
-        return new LadderLine(IntStream.range(NUMBER_ZERO, height)
-                .mapToObj(line -> new Line(countOfUser, rightPointStrategy))
+    public static LadderLines generateLadderLine(int height, int countOfUser, RightPointStrategy rightPointStrategy) {
+        return new LadderLines(IntStream.range(NUMBER_ZERO, height)
+                .mapToObj(line -> new LadderLine(countOfUser, rightPointStrategy))
                 .collect(Collectors.toList()));
     }
