+++ src/test/java/nextstep/ladder/domain/PositionTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.line.Line;
+import nextstep.ladder.domain.line.LadderLine;
 import nextstep.ladder.domain.point.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -10,17 +10,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class PositionTest {
 
-    private Line line;
+    private LadderLine ladderLine;
 
     @BeforeEach
     void setUp() {
-        line = new Line(2, () -> true);
+        ladderLine = new LadderLine(2, () -> true);
     }
 
     @Test
     @DisplayName("포인트 이동 확인")
     void position() {
-        int result = Position.valueOf(line.getPoints(), 0).move(0);
+        int result = Position.valueOf(ladderLine.getPoints(), 0).move(0);
         assertThat(result).isEqualTo(1);
     }
